+++ test/sketch_test.cpp
@@ -1,4 +1,4 @@
-#include "../include/sketch.h"
+#include "../include/l0_sampling/sketch.h"
 #include <chrono>
 #include <gtest/gtest.h>
 #include "util/testing_vector.h"
