+++ test/util/sketch_constructors.h
@@ -1,4 +1,4 @@
-#include "../../include/sketch.h"
+#include "../../include/l0_sampling/sketch.h"
 /*
  * Static functions for creating sketches without a provided memory location
  * used in unit testing.
